refactor: prettyprint is now in the `printers/` folder
